This reverts commit e9cc6adbc30efec912620de397314c57cb60033b.

+++ runtime/base/mutex-inl.h
@@ -129,7 +129,7 @@ class ScopedContentionRecorder {
         owner_tid_(kLogLockContentions ? owner_tid : 0),
         start_nano_time_(kLogLockContentions ? NanoTime() : 0) {
     std::string msg = StringPrintf("Lock contention on %s (owner tid: %llu)",
-                                   mutex->GetName(), (unsigned long long) owner_tid);
+                                   mutex->GetName(), owner_tid);
     ATRACE_BEGIN(msg.c_str());
   }
